fixed mistake on multer

+++ controllers/userController.js
@@ -8,7 +8,7 @@ import path from "path";
 import { DateTime } from "luxon";
 import postsController from "./postsController";
 
-exports.check_friend_status = async function(req, res, next) {
+exports.check_friend_status = async function (req, res, next) {
   const currentUser = req.params.facebookid;
   const friendId = req.params.friendid;
 
@@ -65,7 +65,7 @@ exports.add_friend = asyncHandler(async (req, res, next) => {
   }
 });
 
-exports.remove_friend = async function(req, res, next) {
+exports.remove_friend = async function (req, res, next) {
   const currentUser = req.params.facebookid;
   const friendId = req.params.friendid;
 
@@ -93,7 +93,7 @@ exports.remove_friend = async function(req, res, next) {
   }
 };
 
-exports.accept_friend_request = async function(req, res, next) {
+exports.accept_friend_request = async function (req, res, next) {
   const currentUser = req.params.facebookid;
   const friendId = req.params.friendid;
   try {
@@ -137,7 +137,7 @@ exports.accept_friend_request = async function(req, res, next) {
   }
 };
 
-exports.get_currentuserprofile = async function(req, res) {
+exports.get_currentuserprofile = async function (req, res) {
   try {
     const user = await User.findOne({
       facebook_id: req.params.facebookid
@@ -178,7 +178,7 @@ exports.get_currentuserprofile = async function(req, res) {
   }
 };
 
-exports.get_userprofile = async function(req, res) {
+exports.get_userprofile = async function (req, res) {
   try {
     const otherUser = await User.findOne({ facebook_id: req.params.userid })
       .populate({ path: "country", select: ["country"] })
@@ -240,7 +240,7 @@ exports.get_userprofile = async function(req, res) {
 
 // list of friends
 
-exports.get_listfriends = async function(req, res) {
+exports.get_listfriends = async function (req, res) {
   try {
     const user = await User.findOne({ facebook_id: req.params.facebookid })
       .populate({
@@ -255,7 +255,7 @@ exports.get_listfriends = async function(req, res) {
 };
 // list of other users friends
 
-exports.get_friend_listfriends = async function(req, res) {
+exports.get_friend_listfriends = async function (req, res) {
   try {
     const user = await User.findOne({
       facebook_id: req.params.facebookid
@@ -286,7 +286,7 @@ exports.get_friend_listfriends = async function(req, res) {
 
 // all users
 
-exports.get_users = async function(req, res) {
+exports.get_users = async function (req, res) {
   try {
     const currentUser = await User.findOne(
       {
@@ -312,7 +312,7 @@ exports.get_users = async function(req, res) {
   }
 };
 
-const formatDate = function(dbdate) {
+const formatDate = function (dbdate) {
   if (!dbdate) {
     return undefined;
   }
@@ -321,7 +321,7 @@ const formatDate = function(dbdate) {
 };
 
 // send current profile info
-exports.get_update_profile = async function(req, res) {
+exports.get_update_profile = async function (req, res) {
   try {
     const userprofile = await User.findOne(
       { facebook_id: req.params.facebookid },
@@ -389,7 +389,7 @@ exports.post_update_profile = [
     .isLength({ max: 30 })
     .withMessage("Name length can't exceed 30 characters"),
 
-  async function(req, res) {
+  async function (req, res) {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       return res.json({ errors: errors.array(), profile_content: req.body });
@@ -553,7 +553,7 @@ exports.post_update_profile = [
   }
 ];
 
-exports.post_uploadphoto = async function(req, res) {
+exports.post_uploadphoto = async function (req, res) {
   try {
     console.log("inside  upload photo");
     const userprofilepic = await User.findOne(
@@ -561,14 +561,15 @@ exports.post_uploadphoto = async function(req, res) {
       "profile_pic"
     ).exec();
 
-    console.log("unlink");
-    unlink(path.join(__dirname, "..", userprofilepic.profile_pic), (err) => {
-      if (err) {
-        console.log(err);
-        throw err;
-      }
-    });
-
+    if (!userprofilepic.profile_pic.includes("defaultimage")) {
+      console.log("unlink");
+      unlink(path.join(__dirname, "..", userprofilepic.profile_pic), (err) => {
+        if (err) {
+          console.log(err);
+          throw err;
+        }
+      });
+    }
     await User.findByIdAndUpdate(userprofilepic._id, {
       profile_pic: req.file.path
     }).exec();
@@ -587,7 +588,7 @@ exports.post_search_user = [
     .withMessage("Search keyword must be at least 3 characters")
     .isLength({ max: 20 })
     .withMessage("Search keywords mustn't exceed 30 characters"),
-  async function(req, res) {
+  async function (req, res) {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
       return res.json({ errors: errors.array(), profile_content: req.body });
@@ -623,7 +624,7 @@ exports.post_search_user = [
 
 // browser users near you
 
-exports.get_users_bycountry = async function(req, res) {
+exports.get_users_bycountry = async function (req, res) {
   try {
     const currentUser = await User.findOne(
       {
@@ -653,7 +654,7 @@ exports.get_users_bycountry = async function(req, res) {
   }
 };
 
-const sortBirthdays = function(a, b) {
+const sortBirthdays = function (a, b) {
   if (a.birthday <= b.birthday) {
     return 1;
   } else {
@@ -661,7 +662,7 @@ const sortBirthdays = function(a, b) {
   }
 };
 
-exports.get_birthdays = async function(req, res) {
+exports.get_birthdays = async function (req, res) {
   try {
     const user = await User.findOne(
       {
